Move tab logic from activity to fragment in the Activity feed
