+++ EmbeddedSocialClient/sdk/src/main/java/com/microsoft/embeddedsocial/ui/activity/RecentActivityActivity.java
@@ -5,20 +5,15 @@
 
 package com.microsoft.embeddedsocial.ui.activity;
 
-import android.support.v4.view.PagerAdapter;
-
-import com.microsoft.embeddedsocial.base.GlobalObjectRegistry;
-import com.microsoft.embeddedsocial.sdk.Options;
-import com.microsoft.embeddedsocial.ui.fragment.UserActivityFeedFragment;
+import com.microsoft.embeddedsocial.ui.activity.base.BaseActivity;
+import com.microsoft.embeddedsocial.ui.fragment.FeedViewMenuFragment;
+import com.microsoft.embeddedsocial.ui.fragment.RecentActivityFragmentTabs;
 import com.microsoft.embeddedsocial.sdk.R;
-import com.microsoft.embeddedsocial.ui.activity.base.BaseTabsActivity;
-import com.microsoft.embeddedsocial.ui.fragment.FollowingActivityFeedFragment;
-import com.microsoft.embeddedsocial.ui.util.SimplePagerAdapter;
 
 /**
  * Shows recent activity.
  */
-public class RecentActivityActivity extends BaseTabsActivity {
+public class RecentActivityActivity extends BaseActivity {
 	public static final String NAME = "ActivityFeed";
 
 	public RecentActivityActivity() {
@@ -26,20 +21,10 @@ public class RecentActivityActivity extends BaseTabsActivity {
 	}
 
 	@Override
-	protected PagerAdapter createPagerAdapter() {
-		SimplePagerAdapter.Page userFeed = new SimplePagerAdapter.Page(R.string.es_activity_feed_user, UserActivityFeedFragment::new);
-		SimplePagerAdapter.Page followingFeed = new SimplePagerAdapter.Page(R.string.es_activity_feed_following, FollowingActivityFeedFragment::new);
-
-		SimplePagerAdapter.Page[] pages;
-
-		// Only show the following activity feed if user relations are enabled
-		if (GlobalObjectRegistry.getObject(Options.class).userRelationsEnabled()) {
-			pages = new SimplePagerAdapter.Page[] {userFeed, followingFeed};
-		} else { // user relations are not enabled
-			pages = new SimplePagerAdapter.Page[] {userFeed};
-		}
-
-		return new SimplePagerAdapter(this, getSupportFragmentManager(), pages);
+	protected void setupFragments() {
+		setActivityContent(new RecentActivityFragmentTabs());
+		super.setupFragments();
+		getSupportFragmentManager().beginTransaction().add(new FeedViewMenuFragment(), FeedViewMenuFragment.TAG).commit();
 	}
 
 	@Override
